pyproject.toml
[project]
name = "youtube-2-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cerebras-cloud-sdk>=1.67.0",
    "deepgram-sdk>=5.3.2",
    "fastapi>=0.128.1",
    "fpdf2>=2.8.5",
    "itsdangerous>=2.2.0",
    "langchain>=1.2.9",
    "langchain-openai>=1.1.8",
    "mistralai>=1.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.1.31",
]
